requirements.txt
fastapi
uvicorn
google-cloud-storage
langchain
langchain-google-genai
langchain-community
faiss-cpu
pydantic
langgraph
pypdf